Refactored to use double

+++ include/DataTables/DataTable.h
@@ -6,14 +6,14 @@
 
 class DataTable {
 public:
-    float sumCol(int col);
-    float& at(int row, int col);
-    std::vector<float>& at(int row);
-    void setTableMatrix(std::vector<std::vector<float>> &table);
+    double sumCol(int col);
+    double& at(int row, int col);
+    std::vector<double>& at(int row);
+    void setTableMatrix(std::vector<std::vector<double>> &table);
     int getHeight();
 
 private:
-    std::vector<std::vector<float>> table;
+    std::vector<std::vector<double>> table;
 };
 
 #endif
